Issue #KN-1041 fix: Neo4j & Janus are  Implementations added
