+++ ontology-engine/graph-core_2.12/pom.xml
@@ -30,7 +30,19 @@
         </dependency>
         <dependency>
             <groupId>org.sunbird</groupId>
-            <artifactId>graph-dac</artifactId>
+            <artifactId>graph-janusgraph-impl</artifactId>
+            <exclusions>
+                <exclusion>
+                    <groupId>org.apache.commons</groupId>
+                    <artifactId>commons-lang3</artifactId>
+                </exclusion>
+            </exclusions>
+            <version>1.0-SNAPSHOT</version>
+            <type>jar</type>
+        </dependency>
+        <dependency>
+            <groupId>org.sunbird</groupId>
+            <artifactId>graph-neo4j-impl</artifactId>
             <exclusions>
                 <exclusion>
                     <groupId>org.apache.commons</groupId>
