+++ ontology-engine/graph-neo4j-impl/.gitignore
@@ -1,0 +1,2 @@
+/target
+/bin
