+++ ontology-engine/graph-neo4j-impl/src/main/java/org/sunbird/graph/service/util/GraphQueryGenerationUtil.java
