+++ ontology-engine/pom.xml
@@ -13,7 +13,7 @@
     <modules>
         <module>graph-common</module>
         <module>graph-dac-api</module>
-        <module>graph-dac</module>
+        <module>graph-janusgraph-impl</module>
         <module>graph-core_2.12</module>
         <module>graph-engine_2.12</module>
         <module>parseq</module>
